Support GPT-4
In the process I also implemented major code deduplication in the `conversations` module.
Now I log module used in the `spending` table as well, since it has major impact.
I add a little "(GPT-4)" to the message if the non-default model is used.
However, I didn't get to test it, as apparently I don't have access.