Use a .toml file for system messages
In the process, I also went and renamed identifiers from "system message" to "personality", where it relates to one of the chosen personalities and not just the system message for it. Since system messages can now exist outside of the concept of personalities, it no longer makes sense to call personalities that.