Made all personalities and one-off response commands defined in config
No validation happens now when setting a personality. But if it tries to use an unknown personality, it just uses the default instead. Perhaps that is too silent a failure, but the only way to set a non-existent personality is if a personality is removed or renamed without refreshing the slash commands.