Restrict message link references
For privacy, it can now only be used on messages from the same guild, and on channels the prompter can view.
This necessitated me to add guild IDs to the database, which meant wiping the conversation history. But I never really meant to keep history as long-term as I was anyway.